+++ fedsim/distributed/centralized/training/__init__.py
@@ -1,4 +1,9 @@
-import os
+"""
+Centralized Training
+--------------------
+
+Algorithms for centralized Federated  training.
+"""
 
 from .adabest import AdaBest
 from .fedavg import FedAvg
@@ -7,10 +12,4 @@ from .feddyn import FedDyn
 from .fednova import FedNova
 from .fedprox import FedProx
 
-__all__ = [
-    module[:-3]
-    for module in os.listdir(os.path.dirname(__file__))
-    if module != "__init__.py" and module[-3:] == ".py"
-]
-
-__all__ += ["FedAvg", "FedAvgM", "AdaBest", "FedDyn", "FedNova", "FedProx"]
+__all__ = ["FedAvg", "FedAvgM", "AdaBest", "FedDyn", "FedNova", "FedProx"]