+++ fedsim/distributed/centralized/training/adabest.py
@@ -1,18 +1,13 @@
-r"""This file contains an implementation of the following paper:
-    Title: Title: "Minimizing Client Drift in Federated Learning via Adaptive
-    ---- Bias Estimation"
-    Authors: Farshid Varno, Marzie Saghayi, Laya Rafiee, Sharut Gupta,
-    ---- Stan Matwin, Mohammad Havaei
-    Publication date: [Submitted on 27 Apr 2022 (v1), last revised
-    ---- 23 May 2022 (this version, v2)]
-    Link: https://arxiv.org/abs/2204.13170
+r"""
+AdaBest
+-------
 """
 from functools import partial
 
 import torch
 from torch.nn.utils import parameters_to_vector
 
-from fedsim.local.training.step_closures import default_closure
+from fedsim.local.training.step_closures import default_step_closure
 from fedsim.utils import SerialAggregator
 from fedsim.utils import vector_to_parameters_like
 
@@ -20,6 +15,38 @@ from . import fedavg
 
 
 class AdaBest(fedavg.FedAvg):
+    r"""Implements AdaBest algorithm for centralized FL.
+
+    For further details regarding the algorithm we refer to `AdaBest: Minimizing Client
+    Drift in Federated Learning via Adaptive Bias Estimation`_.
+
+    Args:
+        data_manager (Callable): data manager
+        metric_logger (Callable): a logger object
+        num_clients (int): number of clients
+        sample_scheme (str): mode of sampling clients
+        sample_rate (float): rate of sampling clients
+        model_class (Callable): class for constructing the model
+        epochs (int): number of local epochs
+        loss_fn (Callable): loss function defining local objective
+        batch_size (int): local trianing batch size
+        test_batch_size (int): inference time batch size
+        local_weight_decay (float): weight decay for local optimization
+        slr (float): server learning rate
+        clr (float): client learning rate
+        clr_decay (float): round to round decay for clr (multiplicative)
+        clr_decay_type (str): type of decay for clr (step or cosine)
+        min_clr (float): minimum client learning rate
+        clr_step_size (int): frequency of applying clr_decay
+        device (str): cpu, cuda, or gpu number
+        log_freq (int): frequency of logging
+        mu (float): AdaBest's :math:`\mu` parameter for local regularization
+        beta (float): AdaBest's :math:`\beta` parameter for global regularization
+
+    .. _AdaBest\: Minimizing Client Drift in Federated Learning via Adaptive
+        Bias Estimation: https://arxiv.org/abs/2204.13170
+    """
+
     def __init__(
         self,
         data_manager,
@@ -100,9 +127,7 @@ class AdaBest(fedavg.FedAvg):
         params_init = parameters_to_vector(model.parameters()).detach().clone()
         h = self.read_client(client_id, "h")
         mu_adaptive = (
-            self.mu
-            / len(datasets["train"])
-            * self.read_server("average_sample")
+            self.mu / len(datasets["train"]) * self.read_server("average_sample")
         )
 
         def transform_grads_fn(model):
@@ -115,7 +140,7 @@ class AdaBest(fedavg.FedAvg):
                 p.grad += g_a
 
         step_closure_ = partial(
-            default_closure, transform_grads=transform_grads_fn
+            default_step_closure, transform_grads=transform_grads_fn
         )
         opt_res = super(AdaBest, self).send_to_server(
             client_id,
@@ -134,14 +159,10 @@ class AdaBest(fedavg.FedAvg):
 
         # update local h
         pseudo_grads = (
-            params_init
-            - parameters_to_vector(model.parameters()).detach().clone().data
+            params_init - parameters_to_vector(model.parameters()).detach().clone().data
         )
         t = self.rounds
-        new_h = (
-            1 / (t - self.read_client(client_id, "last_round")) * h
-            + pseudo_grads
-        )
+        new_h = 1 / (t - self.read_client(client_id, "last_round")) * h + pseudo_grads
         self.write_client(client_id, "h", new_h)
         self.write_client(client_id, "last_round", self.rounds)
         return opt_res
@@ -149,9 +170,7 @@ class AdaBest(fedavg.FedAvg):
     def receive_from_client(self, client_id, client_msg, aggregation_results):
         weight = 1
         self.agg(client_id, client_msg, aggregation_results, weight=weight)
-        self.general_agg.add(
-            "avg_m", client_msg["num_samples"] / self.epochs, 1
-        )
+        self.general_agg.add("avg_m", client_msg["num_samples"] / self.epochs, 1)
         self.write_server("average_sample", self.general_agg.get("avg_m"))
 
     def optimize(self, aggregator):