+++ fedsim/distributed/centralized/training/fedavg.py
@@ -1,10 +1,6 @@
-r""" This file contains an implementation of the following paper:
-    Title: "Communication-Efficient Learning of Deep Networks from
-    ---- Decentralized Data"
-    Authors: H. Brendan McMahan, Eider Moore, Daniel Ramage, Seth Hampson,
-    ---- Blaise Agüera y Arcas
-    Publication date: February 17th, 2016
-    Link: https://arxiv.org/abs/1602.05629
+r"""
+FedAvg
+------
 """
 import math
 import sys
@@ -18,12 +14,42 @@ from torch.utils.data import RandomSampler
 
 from fedsim.local.training import local_inference
 from fedsim.local.training import local_train
-from fedsim.local.training.step_closures import default_closure
+from fedsim.local.training.step_closures import default_step_closure
 
 from ..centralized_fl_algorithm import CentralFLAlgorithm
 
 
 class FedAvg(CentralFLAlgorithm):
+    r"""Implements FedAvg algorithm for centralized FL.
+
+    For further details regarding the algorithm we refer to `Communication-Efficient
+    Learning of Deep Networks from Decentralized Data`_.
+
+    Args:
+        data_manager (Callable): data manager
+        metric_logger (Callable): a logger object
+        num_clients (int): number of clients
+        sample_scheme (str): mode of sampling clients
+        sample_rate (float): rate of sampling clients
+        model_class (Callable): class for constructing the model
+        epochs (int): number of local epochs
+        loss_fn (Callable): loss function defining local objective
+        batch_size (int): local trianing batch size
+        test_batch_size (int): inference time batch size
+        local_weight_decay (float): weight decay for local optimization
+        slr (float): server learning rate
+        clr (float): client learning rate
+        clr_decay (float): round to round decay for clr (multiplicative)
+        clr_decay_type (str): type of decay for clr (step or cosine)
+        min_clr (float): minimum client learning rate
+        clr_step_size (int): frequency of applying clr_decay
+        device (str): cpu, cuda, or gpu number
+        log_freq (int): frequency of logging
+
+    .. _Communication-Efficient Learning of Deep Networks from Decentralized
+        Data: https://arxiv.org/abs/1602.05629
+    """
+
     def __init__(
         self,
         data_manager,
@@ -122,7 +148,7 @@ class FedAvg(CentralFLAlgorithm):
         model = ctx["model"]
         optimizer = SGD(model.parameters(), lr=lr, weight_decay=weight_decay)
         # optimize the model locally
-        step_closure_ = default_closure if step_closure is None else step_closure
+        step_closure_ = default_step_closure if step_closure is None else step_closure
         opt_result = local_train(
             model,
             train_loader,