+++ fedsim/distributed/data_management/data_manager.py
@@ -1,3 +1,8 @@
+r"""
+Data Manager
+------------
+
+"""
 import os
 import pickle
 import random