+++ fedsim/distributed/decentralized/compression/__init__.py
@@ -1,0 +1,10 @@
+# """
+
+# Deentralized Compression
+# ------------------------
+
+# There are no decentralized compressions algorithms defined in this version.
+
+# """
+
+# __all__ = []