+++ fedsim/distributed/decentralized/privacy/__init__.py
@@ -1,0 +1,10 @@
+# r"""
+
+# Decentralized Privacy
+# ---------------------
+
+# There are no decentralized privacy algorithms defined in this version.
+
+# """
+
+# __all__ = []