+++ fedsim/scores.py
@@ -1,7 +1,7 @@
-r""" A comnbination of sklearn scores and torch.nn loss functions
-
+r"""
+Fedsim Scores
+-------------
 """
-
 from sklearn import metrics as sk_metrics
 from torch.nn import functional as F