+++ scripts/fedsim_cli.py
@@ -1,8 +1,15 @@
+r"""
+cli Top
+-------
+"""
+
 import os
 
 import click
 
 from .fed_learn import fed_learn
+
+# from .fed_tune import fed_tune
 
 # Enable click
 os.environ["LC_ALL"] = "C.UTF-8"
@@ -15,12 +22,8 @@ def cli():
 
 
 cli.add_command(fed_learn)
-
-
-def main():
-    """main fn"""
-    cli()
+# cli.add_command(fed_tune)
 
 
 if __name__ == "__main__":
-    main()
+    cli()